Remove game_state.rs
