+++ .gitignore
@@ -13,3 +13,6 @@ Cargo.lock
 
 # MSVC Windows builds of rustc generate these, which store debugging information
 *.pdb
+
+# visual studio code
+.vscode*
