+++ src/enums.rs
@@ -1,3 +1,13 @@
+#[derive(Copy, Clone, PartialEq, Debug)]
+pub enum GameState {
+    StartGame = 0,
+    CompleteWord = 1,
+    CompleteAttackWord = 2,
+    Lose = 3,
+    InProgress = 4,
+    Win = 5,
+}
+
 #[derive(Debug)]
 pub enum Message {
     Waiting = 0,
