+++ src/game/mod.rs
@@ -1,4 +1,3 @@
 pub mod game;
-pub mod game_state;
 mod vocab;
 mod word;
