+++ src/lib.rs
@@ -1,6 +1,5 @@
 mod enums;
 mod game;
 
-pub use enums::Message;
+pub use enums::{GameState, Message};
 pub use game::game::{play, Game};
-pub use game::game_state::GameState;
